+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.2.6
+* Added a step function to the univariate data generation module of the `univariate` class.
+* Added the functionality to export the univariate data. An example was added to the univariate notebook in the documentation.
+* Changed univariate data storage to dictionaries instead of arrays.
+
 ## Version 0.2.5
 * Added a check which examples are implemented to the `__init__` function of the `batch` class.